+++ mi/state.py
@@ -159,6 +159,10 @@ class NoteActions:
         res = await self.http.request(Route('POST', '/api/notes/show'), json={"noteId": note_id}, auth=True, lower=True)
         return Note(res, state=self)
 
+    async def get_replies(self, note_id: str, since_id: Optional[str] = None, until_id: Optional[str] = None, limit: int = 10,):
+        res = await self.http.request(Route('POST', '/api/notes/replies'), json={"noteId": note_id, "sinceId": since_id, "untilId": until_id, "limit": limit}, auth=True, lower=True)
+        return [Note(i, state=self) for i in res]
+
 
 class UserAction:
     def __init__(self, http: HTTPClient, loop: asyncio.AbstractEventLoop):